fixed connection issue, added occupation progress wobble
Neutral nodes now show the progress towards capturing as a wobbling circle instead of a population number.
The connection issue was exposed by the AI's ability to act immediately. When you cut a connection to send many links into the target, if the target then created a connection with you, your sent links would change direction.
It also caused a miscounting of used tentacles.